requirements.txt
numpy==2.4.2
pandas==3.0.0
requests==2.32.5
yfinance==1.1.0
python-dotenv==1.2.1
xlwings==0.33.20
reportlab==4.4.10
matplotlib==3.10.8